Add function to DriveTime to also run Intake while driving.
We realized that RunIntake never finished; It was expecting to be
run on a button whileHeld.

The solution here was to add function to DriveTime to also run
the intake.  Another solution would have been to add a parallel
command group 'DriveAndIntake' which did the two things.

This seems to work for now.