No longer invert the rigth side motors - wiring was fixed.
It never made sense to me that we were inverting both sides.
We finally realized that the encoder and motor controller cables
on the right side of the robot were swapped.

The right front motor controller was getting feedback from the
right back encoder and vice versa.

Now that that is fixed, they actually do the right thing.